requirements.txt
langchain==0.2.1
openai==1.30.1
anthropic==0.28.0
python-dotenv==1.0.1
streamlit==1.41.1
loguru==0.7.3
black==24.10.0
isort==5.13.2
transformers==4.41.2
pydantic==2.7.3
torch==2.3.0  # Optional, for advanced ML tasks
tenacity==8.5.0  # For retry mechanism
langchain-openai
langchain-anthropic
langchain-community
langchain-google-genai
google-generativeai
huggingface_hub
